% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtripnum.R
\name{addtripnum}
\alias{addtripnum}
\title{Add a trip number}
\usage{
addtripnum(ans2)
}
\arguments{
\item{ans2}{A \code{\link[base]{data.frame}}. should be a finished SIRI to GTFS table, after \code{\link{check_outlier2}}}
}
\value{
A \code{\link[base]{data.frame}} containing the final result of \code{\link{STG}} with a trip number column added.
}
\description{
Requires the final product (ans2)
adds a unique trip number
}
\details{
Adds a unique trip number during the day for each trip recorded in SIRI,
the function is a part of STG and should not be used on it's own.
}
\section{Warning}{

Do Not use this function on it's own, it is meant to be used only as part of the STG process
}

\references{
Bogin, D., Levy, N. and Ben-Elia E. (2018) \emph{Estimation of Public Transportation Service Reliability Using Big Data and Open Source Tools}
}
\seealso{
\code{\link{STG}}
}
\keyword{internal}
\keyword{misc}
