% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outlier.R
\name{check_outlier}
\alias{check_outlier}
\title{Check for outliers in a SIRI SpatialPointsDataFrame}
\usage{
check_outlier(spSIRI)
}
\arguments{
\item{spSIRI}{A \code{\link[sp]{SpatialPointsDataFrame}}, containing SIRI data.
should be used after \code{\link{SIRItoILTM}} or \code{\link{SIRItoSP}}}
}
\value{
A \code{\link[sp]{SpatialPointsDataFrame}} with an extra column representing the outliers
}
\description{
Add an outlier column to the SIRI data,
Requires the SIRI SpatialPointsDataFrame from \code{\link{SIRItoSP}}
}
\details{
Adds an outlier column to the SIRI data,
the function is a part of STG and should not be used on it's own.
}
\section{Warning}{

Do Not use this function on it's own, it is meant to be used only as part of the STG process
}

\references{
Bogin, D., Levy, N. and Ben-Elia E. (2018) \emph{Estimation of Public Transportation Service Reliability Using Big Data and Open Source Tools}
}
\seealso{
\code{\link[SIRItoGTFS]{STG}}, \code{\link[SIRItoGTFS]{SIRItoILTM}},
\code{\link[SIRItoGTFS]{SIRItoSP}}, \code{\link[sp]{SpatialPointsDataFrame}}
}
\keyword{internal}
\keyword{misc}
