% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outlier2.R
\name{check_outlier2}
\alias{check_outlier2}
\title{Secondary outliers test}
\usage{
check_outlier2(fullans)
}
\arguments{
\item{fullans}{The product of a join between the SIRI data frame after \code{\link{SIRIKeepClosestStop}} and the stop times data frame after \code{\link{organizeStopTimes}}}
}
\value{
A \code{\link[base]{data.frame}} containing only the observations which had 25% of the trip or more
}
\description{
Requires the product of \link[SIRItoGTFS]{SIRIKeepClosestStop}
Checks if there are sufficient observations in the data to analyse the SIRI data.
}
\section{Warning}{

Do Not use this function on it's own, it is meant to be used only as part of the STG process
}

\references{
Bogin, D., Levy, N. and Ben-Elia E. (2018) \emph{Spatial and Temporal Estimation of the Service Reliability of Public Transportation Using Big Data and Open Source Tools}
}
\seealso{
\code{\link{STG}}, \code{\link{SIRIKeepClosestStop}}
}
\keyword{internal}
\keyword{misc}
