% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGTFSil.R
\name{downloadGTFSil}
\alias{downloadGTFSil}
\title{download the Israeli GTFS .zip file}
\usage{
downloadGTFSil(directory = NULL)
}
\arguments{
\item{directory}{If directory remains NULL the function allows you to choose a directory manually via \code{\link[easycsv]{choose_dir}}}
}
\value{
A .zip file in the selected directory, not read into the environment
}
\description{
downloads the Israeli GTFS .zip file into the selected directory with a date added to the files name.
}
\note{
Only useful for Israeli users, other users should seek their local GTFS repository
}
\examples{
\dontrun{
downloadGTFSil()
}
}
\seealso{
\code{\link[easycsv]{choose_dir}}, \url{https://developers.google.com/transit/gtfs/}, \code{\link[utils]{download.file}}
}
\keyword{utilities}
