% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{GTFStrips}
\alias{GTFStrips}
\title{GTFS trips table relevant to SIRI sample}
\format{A data frame with 5172 observations on the following 5 variables:
\describe{
   \item{\code{route_id}}{An ID that uniquely identifies a route.}
   \item{\code{service_id}}{An ID that uniquely identifies a set of dates when service is available for one or more routes. This value is referenced from the calendar table.}
   \item{\code{trip_id}}{An ID that identifies a trip.}
   \item{\code{direction_id}}{A binary value that indicates the direction of travel for a trip. 0 for travel in one direction (e.g. outbound travel), 1 for travel in the opposite direction (e.g. inbound travel).}
   \item{\code{shape_id}}{An ID that defines a shape for the trip. This value is referenced from the shapes table.}
}}
\source{
\url{https://developers.google.com/transit/gtfs/reference/#tripstxt}
}
\usage{
GTFStrips
}
\description{
Subset of the GTFS trips table for Israel for the SIRIsample data.
This is only a small subset of the Israeli GTFS routes table, extracted to be used with the SIRIsample data.
it shows the trips data for the SIRIsample data.
note that GTFS is a global standard and not all operators fill all of the columns.
}
\keyword{datasets}
