% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowmemST2POSIX.R
\name{lowmemST2POSIX}
\alias{lowmemST2POSIX}
\title{convert Stop_times columns to POSIXct on low memory machines}
\usage{
lowmemST2POSIX(column, SIRIref)
}
\arguments{
\item{column}{A \code{\link[base]{vector}} to convert, should be a column in the GTFS Stop_Times table.}

\item{SIRIref}{A SIRI reference table from which to get the date.}
}
\value{
Returns the column as a POSIXct \code{\link[base]{vector}}
}
\description{
A possible replacement function for \code{\link{StopTimes2POSIXct}}.
Is not used by default.
used to convert Date and Time columns in the GTFS Stop_times table to a conformable POSIXct.
}
\details{
use of this function requires editing both the STG function and \code{\link{organizeSIRIdf}} within it to replace all \code{\link{StopTimes2POSIXct}} to lowmemST2POSIX
}
\section{Warning}{

Do Not use this function on it's own, it is meant to be used only as part of the STG process
}

\references{
Bogin, D., Levy, N. and Ben-Elia E. (2018) \emph{Estimation of Public Transportation Service Reliability Using Big Data and Open Source Tools}
}
\seealso{
\code{\link{STG}}, \code{\link{StopTimes2POSIXct}}
}
\keyword{internal}
\keyword{misc}
