% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeStopTimes.R
\name{organizeStopTimes}
\alias{organizeStopTimes}
\title{Organizes the GTFS Stop_times table}
\usage{
organizeStopTimes(Stimes, SIRIdf3.)
}
\arguments{
\item{Stimes}{The product of \code{\link{substoptimes}}, a GTFS stop_times \code{\link[base]{data.frame}} subset.}

\item{SIRIdf3.}{A SIRI \code{\link[base]{data.frame}} after \code{\link{organizeSIRIdf}}.}
}
\value{
A \code{\link[base]{data.frame}} containing columns which can be compared to SIRI's date and time columns.
}
\description{
Requires the product of \code{\link{substoptimes}}.
Prepares the subset of the GTFS stop_times table for comparison with SIRI's real-time data.
}
\section{Warning}{

Do Not use this function on it's own, it is meant to be used only as part of the STG process
}

\references{
Bogin, D., Levy, N. and Ben-Elia E. (2018) \emph{Spatial and Temporal Estimation of the Service Reliability of Public Transportation Using Big Data and Open Source Tools}
}
\seealso{
\code{\link{STG}}, \code{\link{substoptimes}}, \code{\link{organizeSIRIdf}}
}
\keyword{internal}
\keyword{misc}
