% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRIKeepClosestStop.R
\name{SIRIKeepClosestStop}
\alias{SIRIKeepClosestStop}
\title{Keep only the closest stop to each SIRI observation}
\usage{
SIRIKeepClosestStop(SIRIdf)
}
\arguments{
\item{SIRIdf}{A SIRI df with a distance field (SIRIdf4)}
}
\value{
A SIRI \code{\link[base]{data.frame}} with removed duplicates.
}
\description{
Removes duplicates from the SIRI data frame, and keep only the closest stop to each SIRI observation.
requires the data.frame from the product of \code{\link{NearestStop}}.
}
\section{Warning}{

Do Not use this function on it's own, it is meant to be used only as part of the STG process
}

\references{
Bogin, D., Levy, N. and Ben-Elia E. (2018) \emph{Estimation of public transport service reliability using Big Data and open source tools}
}
\seealso{
\code{\link{STG}}
}
\keyword{internal}
\keyword{misc}
