% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRItoGTFS.r
\docType{package}
\name{SIRItoGTFS-package}
\alias{SIRItoGTFS}
\alias{SIRItoGTFS-package}
\title{Compare SIRI real time data with GTFS schedules}
\description{
SIRItoGTFS allows you to compare the actual arrival time of
public transport with the predicted arrival time present in the schedules,
represented with the GTFS tables
}
\details{
The only two functions you're likely to need from \pkg{SIRItoGTFS} are
\code{\link{readGTFS}} and \code{\link{STG}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bogind/SIRItoGTFS}
  \item \url{http://user47094.vs.easily.co.uk/siri/documentation.htm}
  \item \url{https://developers.google.com/transit/gtfs/}
  \item Report bugs at \url{https://github.com/bogind/SIRItoGTFS/issues}
}

}
\author{
\strong{Maintainer}: Dror Bogin \email{dror.bogin@gmail.com}

}
