% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRItoPOSIXct.R
\name{SIRItoPOSIXct}
\alias{SIRItoPOSIXct}
\title{convert SIRI date time columns to POSIXct}
\usage{
SIRItoPOSIXct(column, round = FALSE)
}
\arguments{
\item{column}{the column to reshape into POSIXct}

\item{round}{logical. round to nearest minute?}
}
\value{
the same SIRI column with readable POSIXct values
}
\description{
For internal use in organizeSIRIdf, requires one column from a subset of the entire SIRI dataframe
}
\section{Warning}{

Do Not use this function on it's own, it is meant to be used only as part of the STG process
}

\references{
Bogin, D., Levy, N. and Ben-Elia E. (2018) \emph{Estimation of public transport service reliability using Big Data and open source tools}
}
\seealso{
\code{\link{STG}}, \code{\link{organizeSIRIdf}}
}
\keyword{internal}
\keyword{misc}
