% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StopTimes2POSIXct.R
\name{StopTimes2POSIXct}
\alias{StopTimes2POSIXct}
\title{Converts a Stoptimes column to POSIXct}
\usage{
StopTimes2POSIXct(column, SIRIref)
}
\arguments{
\item{column}{the column to reshape}

\item{SIRIref}{a SIRI reference table from which to get the date}
}
\value{
A \code{\link[base]{vector}} of POSIXct values in "\%Y-\%m-\%d \%H:\%M:\%S" format.
}
\description{
Convert a date time column from the GTFS stop_times table to POSIXct for comparison with SIRI data.
used inside of organizeStoptimes
}
\details{
The function is intended to work with a SIRI reference table containing only date.
the function is a part of \code{\link{organizeStopTimes}} inside \code{\link{STG}} and should not be used on it's own.
it is also interchangeable with \link{lowmemST2POSIX} for advanced user with machines low on memory.
}
\section{Warning}{

Do Not use this function on it's own, it is meant to be used only as part of the STG process
}

\references{
Bogin, D., Levy, N. and Ben-Elia E. (2018) \emph{Estimation of public transport service reliability using Big Data and open source tools}
}
\seealso{
\code{\link{STG}} \code{\link{organizeStopTimes}}, \code{\link{lowmemST2POSIX}}
}
\keyword{internal}
\keyword{misc}
