% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StopsForSIRI.R
\name{StopsForSIRI}
\alias{StopsForSIRI}
\title{subset stops for SIRIdf}
\usage{
StopsForSIRI(SIRI, stops)
}
\arguments{
\item{SIRI}{a SIRIdf after organizeSIRIdf (SIRIdf3)}

\item{stops}{GTFSstops table}
}
\value{
A subset of the GTFS stops table conforming to the bus route being evaluated.
}
\description{
A simple subset of the GTFS stops table to the SIRI data being used in the current analysis.
requires the SIRI data.frame to be after \code{\link{organizeSIRIdf}} and the original GTFS stops table to be used.
}
\section{Warning}{

Do Not use this function on it's own, it is meant to be used only as part of the STG process
}

\references{
Bogin, D., Levy, N. and Ben-Elia E. (2018) \emph{Estimation of public transport service reliability using Big Data and open source tools}
}
\seealso{
\code{\link{STG}}
}
\keyword{internal}
\keyword{misc}
